seq_greedy_bld: Sample the next index using a heap
Previously, the next index was sampled by constructing a cumulative
probability distribution over the values of matrix X, and then doing
inverse CDF sampling. This led to O(n) distribution updates and slowed
down the algorithm considerably.

In this new implementation, we sample the next index deterministically
by storing values and their indices in a max heap. At each time, the
index with the highest value is picked, and its value is decremented.
If its value is 0, it is removed from the heap.

This allows us to perform a O(n) preprocessing step to store the indices
and their values in the heap, O(1) search, and O(logn) value update, and
speeds up the algorithm.