Write psi_appr tests
i. (x < 1): Difference of gsl_sf_psi and util::psi_appr for values in
the sequence 1e-5, 2*1e-5, 3*1e-5, ..., 1 - 1e-5 are tested. Precision
is set to 1e-10.

ii. (x >= 1): Difference of gsl_sf_psi and util::psi_appr for value in
the sequence 1, 1 + 1e-3, 1 + 2*1e-3, ..., 100 - 1e-3 are tested.
Precision is set to 1e-12.