Implement tensor_sums with a CUDA kernel
We now provide our own kernel implementation for summing a tensor along
a given axis. This was needed because we weren't able to map an Eigen
Tensor object to a 3D pitched CUDA memory. This is a problem because it
requires us to store the 3D tensor as a 1D object on the GPU.

In the current implementation, 3D pitched memory addressing is taken
into consideration. Therefore, this function can now be used more easily
in bld algorithms.